+++ ext/Shared/Languages/de_DE.lua
@@ -328,3 +328,5 @@ Language:add(code, "Draw the Points where players can spawn", "Zeichnen Sie die 
 Language:add(code, "The minimum time a bot shoots at one player if in vehicle - recommended minimum 2.5, below this you will have issues.", "Die Mindestzeit, in der ein Bot auf einen Spieler schießt, wenn er sich im Fahrzeug befindet – empfohlenes Minimum 2,5, darunter treten Probleme auf.")
 Language:add(code, "Snipers attack choppers", "Scharfschützen greifen Hubschrauber an")
 Language:add(code, "Bots with sniper-rifels attack choppers", "Bots mit Scharfschützengewehren greifen Hubschrauber an")
+Language:add(code, "Max Bots per vehicle", "Max. Bots pro Fahrzeug")
+Language:add(code, "Maximum number of Bots in a vehicle", "Maximale Anzahl an Bots in einem Fahrzeug")