+++ ext/Shared/Languages/fr_FR.lua
@@ -328,3 +328,5 @@ Language:add(code, "Draw the IDs of the waypoints", "Dessinez les identifiants d
 Language:add(code, "Draw the Points where players can spawn", "Dessinez les points où les joueurs peuvent apparaître")
 Language:add(code, "Snipers attack choppers", "Les tireurs d'élite attaquent les hélicoptères")
 Language:add(code, "Bots with sniper-rifels attack choppers", "Des bots avec des fusils de sniper attaquent des hélicoptères")
+Language:add(code, "Max Bots per vehicle", "Max Bots par véhicule")
+Language:add(code, "Maximum number of Bots in a vehicle", "Nombre maximum de Bots dans un véhicule")